backend_django/requirements.txt
# Django Core
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# File Storage & Media
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# API Documentation
drf-spectacular==0.27.0

# Environment
python-dotenv==1.0.0

# Utilities
django-filter==23.5
django-crispy-forms==2.1
crispy-bootstrap5==2.0.0

# Celery for async tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# AI/ML
openai==1.6.1
anthropic==0.8.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0

# Email
django-ses==3.5.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.39.2

# Rate Limiting
django-ratelimit==4.1.0

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Timezone
pytz==2023.3
